requirements.txt
addict==2.4.0
attr==0.3.2
attrs==21.4.0
bleach==5.0.0
bpy==3.4.0
crc32c==2.3
cupy==11.5.0
cupy_cuda11x==11.0.0
Cython==0.29.32
einops==0.4.1
executing==1.2.0
gym==0.19.0
h5py==3.7.0
imageio==2.19.3
ipython==8.10.0
lmdb==1.3.0
lxml==4.9.2
MarkupSafe==2.1.1
mathutils==3.3.0
matplotlib==3.5.2
mc==0.0
mise==1.0
networkx==2.8.5
numba==0.56.4
numpy==1.22.3
omegaconf==1.4.1
open3d==0.12.0
pandas==1.5.3
Pillow==9.4.0
plyfile==0.7.4
psutil==5.9.1
pycuda==2022.2.2
pykdtree==1.3.5
pynput==1.7.6
pynvml==11.4.1
pytest==7.1.2
python-dotenv==0.21.1
pytorch3d==0.3.0
pytorch_lightning==0.7.1
PyTurboJPEG==1.6.7
PyYAML==6.0
regex==2022.10.31
requests==2.27.1
rich==13.3.1
rospy==1.15.14
sapien==2.0.0.dev20220701
scikit_image==0.18.3
scikit_learn==1.2.1
scipy==1.10.1
seaborn==0.11.2
setuptools==61.2.0
Shapely==1.8.2
six==1.16.0
skimage==0.0
sorcery==0.2.2
tensorboard==2.9.1
tensorboardX==2.6
termcolor==1.1.0
torch_scatter==2.1.0
torchmetrics==0.4.1
torchvision==0.12.0
tqdm==4.64.1
transforms3d==0.4.1
trimesh==3.13.0
wandb==0.13.2
yapf==0.32.0
hydra-core==0.11.3
